Refactor lexplore module into logical submodules
Split the monolithic 1350-line lexplore.rs into focused submodules:

- lexplore/loader.rs (805 lines): File loading, parsing, tokenizing
  - ElementType, DocumentType, Parser enums
  - ElementLoader, ParsedElement, ParsedTokens structs
  - Lexplore main interface with all loading methods
  - parse_with_parser, tokenize_with_parser utilities

- lexplore/extraction.rs (107 lines): AST extraction helpers
  - get_first_* functions for element extraction
  - paragraph_text_* assertion helpers
  - documents_match, content_items_match comparison utilities

- lexplore.rs (451 lines): Module organization and tests
  - Module declarations and public re-exports
  - All 45 tests (maintained test coverage)

Benefits:
- Clear separation of concerns (loading vs extraction)
- Each module has a single responsibility
- Easier to navigate and maintain
- All 361 tests passing
